// src/api/generated/schemas/postClothingItem.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ClothingFit API
 * ClothingFit 백엔드 API 문서
 * OpenAPI spec version: 1.0.0
 */
import type { PostClothingItemType } from './postClothingItemType';

export interface PostClothingItem {
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  brand: string | null;
  name: string;
  /** @nullable */
  size: string | null;
  /** @nullable */
  link: string | null;
  type: PostClothingItemType;
}